requirements.txt
# Core dependencies
openai>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# SAM (Segment Anything)
torch>=2.0.0
torchvision>=0.15.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# RAG dependencies
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
rich>=13.0.0
